Moved bitmap to STD library
