+++ SOURCE/STD/BITMAP.h
